functions/xml-to-json-converter/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
cloudevents==1.*
cloud-sql-python-connector==1.*
psycopg2-binary==2.*